fix environment iteration that made over the resolver imports instead of the engine's origins. add comments
