+++ analyzer/checker/src/steps.rs
@@ -1,3 +1,19 @@
+use parser::err::ParseError;
+use crate::importer::ImportError;
+
 pub mod collect;
-mod lib;
 mod resolve;
+
+
+#[derive(Debug, PartialEq)]
+pub enum GatherError {
+    Import(ImportError),
+    Parse(Vec<ParseError>),
+    Other(String),
+}
+
+impl From<ImportError> for GatherError {
+    fn from(err: ImportError) -> Self {
+        GatherError::Import(err)
+    }
+}
