--- analyzer/checker/src/steps/lib.rs
@@ -1,15 +1,0 @@
-use crate::import::ImportError;
-use parser::err::ParseError;
-
-#[derive(Debug, PartialEq)]
-pub enum GatherError {
-    Import(ImportError),
-    Parse(Vec<ParseError>),
-    Other(String),
-}
-
-impl From<ImportError> for GatherError {
-    fn from(err: ImportError) -> Self {
-        GatherError::Import(err)
-    }
-}
